Fix build for testing, all tx tests are passed

+++ src/main/Config.h
@@ -165,7 +165,7 @@ class Config : public std::enable_shared_from_this<Config>
     uint64_t adminRoleID;
 
     std::string BASE_EXCHANGE_NAME;
-    uint32_t TX_EXPIRATION_PERIOD;
+    int64_t TX_EXPIRATION_PERIOD;
     int64 TX_EXPIRATION_PERIOD_WINDOW = 60*60;
     
     uint64 MAX_INVOICES_FOR_RECEIVER_ACCOUNT = 20;
